% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_regions_from_bed.R
\name{annotate_regions_from_bed}
\alias{annotate_regions_from_bed}
\title{Annotate regions according to a provided bed annotation file}
\usage{
annotate_regions_from_bed(object, bed_path)
}
\arguments{
\item{object}{TChroNetNetwork object}

\item{bed_path}{Path to a bed file containing regions and annotations}
}
\value{
A TChroNetNetwork object
}
\description{
Annotate regions according to a provided bed annotation file
}
