% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_TChroNetSeries_object.R
\name{build_TChroNetSeries_object}
\alias{build_TChroNetSeries_object}
\title{Create the TChroNetSeries object}
\usage{
build_TChroNetSeries_object(
  edge_files,
  method = NULL,
  resolutions_list = seq(0.1, 0.9, 0.1),
  seed = 123
)
}
\arguments{
\item{edge_files}{List of hd5 files to build the network}

\item{method}{Method for community investigation (c("Leiden", "Louvain"))}

\item{resolutions_list}{List of resolutions to be investigated}

\item{seed}{Set a seed for computational reproducibility}
}
\value{
A TChroNetSeries object
}
\description{
Create the TChroNetSeries object
}
