% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_TChroNetSeries_object.R
\name{.add_weighted_self_loops_iso}
\alias{.add_weighted_self_loops_iso}
\title{Create the TChroNetSeries object}
\usage{
.add_weighted_self_loops_iso(G, loop_weight = 1)
}
\arguments{
\item{edge_files}{List of parquet files to build the network}

\item{matrix_path}{Path to the log2 normalized count matrix used in tchronetPy}

\item{method}{Method for community investigation (c("Leiden", "Louvain"))}

\item{resolutions_list}{List of resolutions to be investigated}

\item{run_cd}{Boolean value for running community detection evaluation. default = FALSE}

\item{transitivity}{Boolean value for evaluating transitivity value for each threshold. default = FALSE}

\item{seed}{Set a seed for computational reproducibility}
}
\value{
A TChroNetSeries object
}
\description{
Create the TChroNetSeries object
}
