% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_best_resolution.R
\name{find_best_resolution}
\alias{find_best_resolution}
\title{Find the best resolution paramenter according to evaluated modularities}
\usage{
find_best_resolution(object)
}
\arguments{
\item{object}{TChroNetNetwork object}
}
\value{
An object of class TChroNetNetwork
}
\description{
Find the best resolution paramenter according to evaluated modularities
}
