% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_position_stacked.R
\name{genomic_position_stackbar}
\alias{genomic_position_stackbar}
\title{Evalaute the annotations for each investigated region}
\usage{
genomic_position_stackbar(object, resolution = NULL)
}
\arguments{
\item{object}{TChroNetNetwork object}

\item{resolutions}{Set the resolution parameter for plotting trends. If not specified the function the resolution parameter stored in the object (deafault = NULL)}
}
\value{
A dataframe containing annotations for each regions
}
\description{
Evalaute the annotations for each investigated region
}
