% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interrogate_GREAT_regions.R
\name{interrogate_GREAT_regions}
\alias{interrogate_GREAT_regions}
\title{Interrogate GREAT and retrieve a list of target genes}
\usage{
interrogate_GREAT_regions(gr, comm_name, species)
}
\arguments{
\item{gr}{GRanges object}

\item{comm_name}{Name of the community}

\item{species}{Name of the genome to be used}
}
\value{
A list of target genes
}
\description{
Interrogate GREAT and retrieve a list of target genes
}
