% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifted_network_coordinates.R
\name{lift_network_coordinates}
\alias{lift_network_coordinates}
\title{Lift the coordinates from a genome to another}
\usage{
lift_network_coordinates(
  object,
  chain_path,
  source_build = "hg19",
  target_build = "hg38"
)
}
\arguments{
\item{object}{TChroNetNetwork object}

\item{chain_path}{Path to a chain file for liftover}

\item{source_build}{Name of source genome (default = 'hg19')}

\item{target_build}{Name of target genome (default = 'hg38')}
}
\value{
A TChroNetNetwork object
}
\description{
Lift the coordinates from a genome to another
}
