% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_TChroNetNetwork_object.R
\name{make_TChroNetNetwork_obj}
\alias{make_TChroNetNetwork_obj}
\title{Create a NetworkAnalysis object from an HDF5 network file}
\usage{
make_TChroNetNetwork_obj(network_path, matrix_path)
}
\arguments{
\item{network_path}{Path to the HDF5 file}

\item{matrix_path}{Path to the tsv normalized counts matrix}
}
\value{
An object of class NetworkAnalysis
}
\description{
Create a NetworkAnalysis object from an HDF5 network file
}
