% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_community_sankey.R
\name{plot_community_sankey}
\alias{plot_community_sankey}
\title{Plot the sankey plot of communities at increased resolution for a fixed threshold}
\usage{
plot_community_sankey(object, threshold = NULL, output_path = NULL)
}
\arguments{
\item{object}{TCrhoNetSeries or TCrhoNetNetwork object}

\item{threshold}{Set the threshold to be plotted if TCrhoNetSeries object was passed. default = NULL}
}
\value{
A ggplot2 object
}
\description{
Plot the sankey plot of communities at increased resolution for a fixed threshold
}
