% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{Plot the density at each threshold}
\usage{
plot_density(object)
}
\arguments{
\item{object}{TCrhoNetSeries object}
}
\value{
A ggplot2 object
}
\description{
Plot the density at each threshold
}
