% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrichments_tfs.R
\name{plot_enrichments_tfs}
\alias{plot_enrichments_tfs}
\title{Annotate regions according to a provided bed annotation file}
\usage{
plot_enrichments_tfs(object, negLog10Padj_th = 4, log2enr_th = 1, top_n = 10)
}
\arguments{
\item{object}{TChroNetNetwork object}

\item{negLog10Padj_th}{negative log 10 p-value threshold. defult = 4.0}

\item{log2enr_th}{log 2 enrichment threshold. defult = 1.0}

\item{top_n}{number of most enriched transcription factors for each community to be plotted}
}
\value{
e ggplot2 object
}
\description{
Annotate regions according to a provided bed annotation file
}
