% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metrics.R
\name{plot_metrics}
\alias{plot_metrics}
\title{Plot the transitivity at each threshold}
\usage{
plot_metrics(object, metric = "relative_lcc")
}
\arguments{
\item{object}{TCrhoNetSeries object}
}
\value{
A ggplot2 object
}
\description{
Plot the transitivity at each threshold
}
