% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_modularity.R
\name{plot_modularity}
\alias{plot_modularity}
\title{Plot the modularity. If TCrhoNetSeries object is provided the plot represents all the available threshold}
\usage{
plot_modularity(object)
}
\arguments{
\item{object}{TChroNetNetwork object}
}
\value{
A ggplot2 object
}
\description{
Plot the modularity. If TCrhoNetSeries object is provided the plot represents all the available threshold
}
