% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sankey_fixed_resolution.R
\name{plot_sankey_fixed_resolution}
\alias{plot_sankey_fixed_resolution}
\title{Plot the sankey plot of communities at increased threshod for a fixed resolution}
\usage{
plot_sankey_fixed_resolution(
  object,
  resolution = NULL,
  thresholds = NULL,
  output_path = NULL
)
}
\arguments{
\item{object}{TCrhoNetSeries object}

\item{resolution}{Fixed resolution to compare}

\item{thresholds}{List of threshold to compare. If not provided it uses all the sotred thresholds}
}
\value{
A ggplot2 object
}
\description{
Plot the sankey plot of communities at increased threshod for a fixed resolution
}
