% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stacked_annotation.R
\name{plot_stacked_annotation}
\alias{plot_stacked_annotation}
\title{Plot stacked barplot of annotations per communities}
\usage{
plot_stacked_annotation(object, resolution = NULL)
}
\arguments{
\item{object}{TChroNetNetwork object}

\item{resolutions}{Set the resolution parameter for plotting trends. If not specified the function the resolution parameter stored in the object (deafault = NULL)}
}
\value{
A ggplot2 object
}
\description{
Plot stacked barplot of annotations per communities
}
