% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_transitivity.R
\name{plot_transitivity}
\alias{plot_transitivity}
\title{Plot the transitivity at each threshold}
\usage{
plot_transitivity(object)
}
\arguments{
\item{object}{TCrhoNetSeries object}
}
\value{
A ggplot2 object
}
\description{
Plot the transitivity at each threshold
}
