% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_community_biology.R
\name{profile_community_biology}
\alias{profile_community_biology}
\title{Profile Community Biology}
\usage{
profile_community_biology(object, genome, phast, txdb, resolution)
}
\arguments{
\item{object}{TChroNetNetwork object}

\item{genome}{A BSgenome object (e.g., BSgenome.Hsapiens.UCSC.hg38).}

\item{phast}{A GScores or GenomicScores object (e.g., phastCons100way.UCSC.hg38).}

\item{txdb}{A TxDb object for TSS calculation (e.g., TxDb.Hsapiens.UCSC.hg38.knownGene).}

\item{resolution}{The resolution value used for community detection}
}
\value{
A dataframe containing statistics for each genomic region
}
\description{
Profile Community Biology
}
