% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_GREAT_analysis.R
\name{run_GREAT_analysis}
\alias{run_GREAT_analysis}
\title{Find taret genes for each community accordig GREAT analysis}
\usage{
run_GREAT_analysis(object, resolution = NULL, genome = "hg19")
}
\arguments{
\item{object}{TChroNetNetwork object}

\item{genome}{Genome to investigate (default = 'hg19')}

\item{resolutions}{Set the resolution parameter for plotting trends. If not specified the function the resolution parameter stored in the object (deafault = NULL)}
}
\value{
A TChroNetNetwork object
}
\description{
Find taret genes for each community accordig GREAT analysis
}
