% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_monalisa_enrichment.R
\name{run_monalisa_enrichment}
\alias{run_monalisa_enrichment}
\title{Annotate regions according to a provided bed annotation file}
\usage{
run_monalisa_enrichment(
  object,
  pwms,
  resolution = NULL,
  genome = NULL,
  num_cores = 1
)
}
\arguments{
\item{object}{TChroNetNetwork object}

\item{pwms}{PWMatrixList with motifs for which to calculate enrichments.}

\item{resolution}{Resolution value to calcualte communities enrichemnt, if NULL the default resolution is used if calculated. default = NULL}

\item{genome}{specify the genome to use, choose among "hg19", "hg38" , "mm10". default = NULL}

\item{num_cores}{Number of cores to be used in monaLisa analysis. defult = 1}
}
\value{
A TChroNetNetwork object
}
\description{
Annotate regions according to a provided bed annotation file
}
