% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{TChroNetR-package}
\alias{TChroNetR}
\alias{TChroNetR-package}
\title{TChroNetR: T-ChroNet Analysis Toolkit}
\description{
This package allows the network analysis of Chromatin Networks.
}
\author{
\strong{Maintainer}: Stefano Di Giovenale \email{stefano.digiovenale@ifo.it}

}
\keyword{internal}
