% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{TCrhoNetNetwork-class}
\alias{TCrhoNetNetwork-class}
\title{TCrhoNetNetwork class}
\description{
TCrhoNetNetwork class
}
\section{Slots}{

\describe{
\item{\code{graph}}{igraph object}

\item{\code{matrix}}{Matrix of normalized log2 counts used for inferring the netowrk}

\item{\code{genomicRegions}}{GRanges object of genomic regions}

\item{\code{lifted_coords}}{Lifted genome coordinates from original genome to raget genome}

\item{\code{clusters}}{Data frame of cluster memberships for different resolution parameters}

\item{\code{modularity}}{Data frame of modularity values for each resolution}

\item{\code{metadata}}{Data frame containing the metadata information of netowrk (number of edges, number of nodes, density)}

\item{\code{resolution}}{Value of resolution to be used. It stores the last resolution value used or the best resolution value}

\item{\code{annotations}}{Annotation table from BED overlaps}

\item{\code{GREAT_targets}}{List of GREAT target genes}
}}

