% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_TChroNetSeries_object.R
\name{build_TChroNetSeries_object}
\alias{build_TChroNetSeries_object}
\title{Create the TChroNetSeries object}
\usage{
build_TChroNetSeries_object(
  output_dir,
  matrix_path,
  method = c("Leiden", "Louvain"),
  resolutions_list = seq(0.1, 0.9, 0.1),
  run_cd = FALSE,
  transitivity = FALSE,
  min_size = 100,
  merged_id = 1000,
  seed = 123
)
}
\arguments{
\item{output_dir}{Directory containing TChroNetPy results}

\item{matrix_path}{Path to the log2 normalized count matrix used in tchronetPy}

\item{method}{Method for community investigation (c("Leiden", "Louvain"))}

\item{resolutions_list}{List of resolutions to be investigated}

\item{run_cd}{Boolean value for running community detection evaluation. default = FALSE}

\item{transitivity}{Boolean value for evaluating transitivity value for each threshold. default = FALSE}

\item{min_size}{Size of the smallest community to consider}

\item{merged_id}{The community id given to all the nodes belonging to communities smaller than min_size}

\item{seed}{Set a seed for computational reproducibility}
}
\value{
A TChroNetSeries object
}
\description{
Create the TChroNetSeries object
}
