% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_TCrhoNetNetwork_from_series.R
\name{build_TCrhoNetNetwork_from_series}
\alias{build_TCrhoNetNetwork_from_series}
\title{Convert a TCrhoNetSeries into a TCrhoNetNetwork}
\usage{
build_TCrhoNetNetwork_from_series(
  series_object,
  threshold,
  matrix_path,
  verbose = TRUE
)
}
\arguments{
\item{threshold}{Choosen correlation threshold. If not provided it uses the best threshold parameter stored in the TCrhoNetSeries object}

\item{matrix_path}{Path of the normalized log2 matrix used for correlation evaluation}

\item{verbose}{(default=TRUE)}

\item{object}{TCrhoNetSeries object}
}
\value{
A TCrhoNetNetwork object
}
\description{
Convert a TCrhoNetSeries into a TCrhoNetNetwork
}
