% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_membership_nodes.R
\name{compute_membership_nodes}
\alias{compute_membership_nodes}
\title{Calcualte communities in network}
\usage{
compute_membership_nodes(
  object,
  resolutions = 1,
  method = c("Leiden", "Louvain"),
  niter = 2,
  min_size = 100,
  merged_id = 1000,
  seed = 1234
)
}
\arguments{
\item{object}{TChroNetNetwork object}

\item{resolutions}{Resolution parameter it can be a unique resolution or a list of resolutions}

\item{method}{Choose a 'Leiden' or 'Louvain' method for community detection}

\item{niter}{Number of Leiden iterations that the algorithm should be run for (default=2)}

\item{seed}{Set a seed for computational reproducibility}
}
\value{
An object of class TChroNetNetwork
}
\description{
Calcualte communities in network
}
