% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_best_th.R
\name{find_best_th}
\alias{find_best_th}
\title{Find the best threshold value according to the transitivity values at increased thresholds}
\usage{
find_best_th(object, parameter = "relative_lcc")
}
\arguments{
\item{object}{TCrhoNetSeries object}
}
\value{
A TCrhoNetSeries object
}
\description{
Find the best threshold value according to the transitivity values at increased thresholds
}
