% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_randindex_map.R,
%   R/plot_randindex_map_old.R
\name{plot_randindex_map}
\alias{plot_randindex_map}
\title{Plot the random index map in which communities are compared among neighbors at increased threshold and resolution.}
\usage{
plot_randindex_map(object, output_path = NULL, color_limits = c(0, 1))

plot_randindex_map(object, output_path = NULL, color_limits = c(0, 1))
}
\arguments{
\item{object}{TCrhoNetSeries object}

\item{color_limits}{List of min and max limits for color map (default = c(0.0,1))}
}
\value{
A ggplot2 object

A ggplot2 object
}
\description{
Plot the random index map in which communities are compared among neighbors at increased threshold and resolution.

Plot the random index map in which communities are compared among neighbors at increased threshold and resolution.
}
