% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trends_zscore.R
\name{plot_trends_zscore}
\alias{plot_trends_zscore}
\title{Plot communities trends at a specified resolution}
\usage{
plot_trends_zscore(object, resolution = NULL, custom_ylim = c(-2, 2))
}
\arguments{
\item{object}{TChroNetNetwork object}

\item{custom_ylim}{list of min and max value for plotted trends (deafult = c(-2,2))}

\item{resolutions}{Set the resolution parameter for plotting trends. If not specified the function the resolution parameter stored in the object (deafault = NULL)}
}
\value{
A ggplot2 object
}
\description{
Plot communities trends at a specified resolution
}
