% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{TCrhoNetSeries-class}
\alias{TCrhoNetSeries-class}
\title{TCrhoNetTCrhoNetSeriesNetwork class}
\description{
TCrhoNetTCrhoNetSeriesNetwork class
}
\section{Slots}{

\describe{
\item{\code{edge_files}}{List of edge hd5 files}

\item{\code{thresholds}}{List of all the threshold stored}

\item{\code{metrics}}{Data frame containing the metrics for each threshold}

\item{\code{communities}}{List of communities found for each threshold}

\item{\code{lifted_coords}}{Lifted genome coordinates from original genome to raget genome}

\item{\code{modularity}}{Data frame of modularity values for each resolution and each threshold}

\item{\code{metadata}}{Data frame containing the metadata information of netowrk (number of edges, number of nodes, density)}

\item{\code{components}}{Data frame of detected components at increased resolution}

\item{\code{method}}{Method used for investigation of communities}

\item{\code{best_th}}{Value of best threshold value}
}}

